pass-through: provide leading zero for vdevfn
This is consistent with the format used elsewhere.

At this time this inconsistency manifests in the output of pci-list
for functions that have been hot-plugged.

While the pci-list output could be unified inside xm or xend by,
for example parsing the string as an integer and then formating it
as a string, there seems to be no disadvantage in making the
string representation used consistent.

Cc: Masaki Kanno <kanno.masaki@jp.fujitsu.com>
Cc: Dexuan Cui <dexuan.cui@intel.com>
Signed-off-by: Simon Horman <horms@verge.net.au>